chore: 删除 Quartz 数据库表结构
- 移除了 Quartz相关的数据库表结构定义- 包括任务详细信息表、触发器详细信息表、简单触发器表、Cron 触发器表、Blob 类型触发器表、日历信息表、暂停的触发器表、已触发的触发器表、调度器状态表、悲观锁信息表和同步机制的行锁表

+++ softcostEstimator-admin/src/main/java/com/softcostEstimator/evaluate/domain/FunctionPointAnalysis.java
@@ -48,11 +48,11 @@ public class FunctionPointAnalysis extends BaseEntity
 
     /** TCF */
     @Excel(name = "TCF")
-    private Long TCF;
+    private Float TCF;
 
     /** AFP */
     @Excel(name = "AFP")
-    private Long AFP;
+    private Float AFP;
 
     public void setProjectId(Long projectId) 
     {
@@ -126,21 +126,21 @@ public class FunctionPointAnalysis extends BaseEntity
     {
         return GSC;
     }
-    public void setTCF(Long TCF) 
+    public void setTCF(float TCF)
     {
         this.TCF = TCF;
     }
 
-    public Long getTCF() 
+    public Float getTCF()
     {
         return TCF;
     }
-    public void setAFP(Long AFP) 
+    public void setAFP(float AFP)
     {
         this.AFP = AFP;
     }
 
-    public Long getAFP() 
+    public Float getAFP()
     {
         return AFP;
     }
